+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * FileUtil增加rename方法
 * 增加Copier接口抽象拷贝
 * 增加FileCopier强化文件拷贝
+* ZipUtil增加多个流的方法（issue#IE5ZC@osc）
 
 ### Bug修复
 * 修复Entity.parseEntity方法中获取表名逻辑的问题（感谢@【北京】游弋苍茫）